services/users/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
httpx>=0.27,<1.0
pydantic>=2.6,<3.0
opentelemetry-api>=1.24,<2.0
opentelemetry-sdk>=1.24,<2.0
opentelemetry-exporter-otlp>=1.24,<2.0
opentelemetry-instrumentation-fastapi>=0.45b0,<1.0
opentelemetry-instrumentation-httpx>=0.45b0,<1.0
opentelemetry-exporter-otlp-proto-grpc
python-json-logger>=2.0.7,<3.0
asyncpg>=0.29,<1.0
